Linter

+++ packages/react-native-worklets/Common/cpp/worklets/Public/AsyncQueue.h
@@ -10,7 +10,7 @@ class AsyncQueue : public facebook::jsi::NativeState {
 
   virtual void push(std::function<void()> &&job) = 0;
   virtual void pushPriority(std::function<void()> &&job) {}
-  virtual void pushTimeout(std::function<void()> &&job, long long delay) {}
+  virtual void pushTimeout(std::function<void()> &&job, int64_t delay) {}
 };
 
 } // namespace worklets
